+++ org.mmisw.orrportal/src/org/mmisw/orrportal/gwt/server/OrrConfig.java
@@ -1,14 +1,34 @@
 package org.mmisw.orrportal.gwt.server;
 
 import com.typesafe.config.Config;
+import com.typesafe.config.ConfigFactory;
+import org.apache.commons.logging.Log;
+import org.apache.commons.logging.LogFactory;
+
+import java.io.File;
+import java.util.ServiceConfigurationError;
 
 /**
  * ORR runtime configuration.
  */
 public class OrrConfig {
 
-  public static OrrConfig load(Config cfg) {
+  public static OrrConfig init() {
+    Config baseAppConfig = ConfigFactory.load();
+    String configFilename = baseAppConfig.getString("configFile");
+    if (configFilename == null) {
+      throw new ServiceConfigurationError("Could not retrieve configuration parameter: configFile");
+    }
+    log.info("Loading configuration from " + configFilename);
+    File configFile = new File(configFilename);
+    if (!configFile.canRead()) {
+      throw new ServiceConfigurationError("Could not read configuration file: " + configFile);
+    }
+
+    Config cfg = ConfigFactory.parseFile(configFile);
+
     orrConfig = new OrrConfig(cfg);
+    log.info("Loaded OrrConfig: " + orrConfig);
     return orrConfig;
   }
 
@@ -17,25 +37,77 @@ public class OrrConfig {
   }
 
 
-  public final String workspace;
+  public final File workspaceDir;
 
-  public final boolean useAgraph;
+  /** where the "pre-loaded" files are stored: (a pre-loaded file is one that the user uploads
+   * when starting the process of submitting an ontology to the repository) */
+  public final File preUploadsDir;
+
+  public final File voc2rdfDir;
+  public final File previewDir;
+  public final File resourceDir;
+
   public final String agraphSparql;
+
+  public final String emailUsername;
+  public final String emailPassword;
 
   @Override
   public String toString() {
-    return "OrrConfig{" +
-        "workspace='" + workspace + '\'' +
-        ", agraphSparql='" + agraphSparql + '\'' +
+    return "OrrConfig{\n" +
+        "  workspaceDir =" + workspaceDir + '\n' +
+        "  preUploadsDir=" + preUploadsDir + '\n' +
+        "  voc2rdfDir   =" + voc2rdfDir + '\n' +
+        "  previewDir   =" + previewDir + '\n' +
+        "  resourceDir  =" + resourceDir + '\n' +
+        "  agraphSparql =" + agraphSparql + '\n' +
+        "  emailUsername=" + emailUsername +
         '}';
   }
 
   private OrrConfig(Config cfg) {
-    workspace = cfg.getString("workspace");
+    workspaceDir = prepareDir(cfg.getString("workspace"));
 
-    useAgraph = cfg.hasPath("agraph");
-    agraphSparql = useAgraph ? cfg.getString("agraph.sparql") : null;
+    preUploadsDir = prepareDir(workspaceDir, "ontmd/preuploads");
+    voc2rdfDir    = prepareDir(workspaceDir, "ontmd/preuploads/voc2rdf");
+    previewDir    = prepareDir(workspaceDir, "ontmd/previews");
+    resourceDir   = prepareDir(workspaceDir, "ontmd/resources");
+
+    if(cfg.hasPath("email")) {
+      emailUsername = cfg.getString("email.username");
+      emailPassword = cfg.getString("email.password");
+    }
+    else {
+      emailUsername = emailPassword = null;
+    }
+
+    agraphSparql = cfg.hasPath("agraph.sparql") ? cfg.getString("agraph.sparql") : null;
+  }
+
+  private static File prepareDir(String dirName) {
+    return prepareDir(new File(dirName));
+  }
+
+  private static File prepareDir(File parentDir, String subdirName) {
+    return prepareDir(new File(parentDir, subdirName));
+  }
+
+  private static File prepareDir(File dirFile) {
+    if (dirFile.exists()) {
+      if (!dirFile.isDirectory()) {
+        throw new RuntimeException(dirFile + " already exists but is not a directory");
+      }
+    }
+    else {
+      if (!dirFile.mkdirs()) {
+        throw new RuntimeException("could not create directory: " + dirFile);
+      }
+    }
+
+    return dirFile;
   }
 
   private static OrrConfig orrConfig;
+
+  private static final Log log = LogFactory.getLog(OrrConfig.class);
 }