+++ ChangeLog.md
@@ -3,6 +3,7 @@
 * 2015-11-10: orrportal 2.5.4
   - preparations for #353 "streamline installation of ORR system"
   	- add template.mmiorr.conf and typesafe config library
+  	- add application.conf and OrrConfig
   
 * 2015-11-06: ont\&orrportal 2.5.3
   - refactoring in orrportal to facilitate possible reimplementation of interface against repository